requirements.txt
# ============================================================
# QueryKeys Polymarket Bot — Dependencies
# Python >= 3.11
# ============================================================

# ---- Polymarket Official Client ----------------------------
py-clob-client>=0.17.0

# ---- Async HTTP & WebSocket --------------------------------
aiohttp>=3.9.0
aiofiles>=23.2.1
websockets>=12.0

# ---- Database ----------------------------------------------
sqlalchemy>=2.0.27
aiosqlite>=0.19.0
asyncpg>=0.29.0            # PostgreSQL async driver
alembic>=1.13.0

# ---- Configuration & Env -----------------------------------
pydantic>=2.6.0
pydantic-settings>=2.2.0
python-dotenv>=1.0.1
pyyaml>=6.0.1

# ---- Logging -----------------------------------------------
structlog>=24.1.0

# ---- ML Models ---------------------------------------------
lightgbm>=4.3.0
xgboost>=2.0.3
catboost>=1.2.3
scikit-learn>=1.4.0

# ---- Data Science ------------------------------------------
numpy>=1.26.4
pandas>=2.2.0
scipy>=1.12.0

# ---- LLM / Anthropic Claude --------------------------------
anthropic>=0.25.0          # Claude API with prompt caching

# ---- Sentiment Analysis ------------------------------------
vaderSentiment>=3.3.2

# ---- Dashboard ---------------------------------------------
streamlit>=1.32.0
plotly>=5.20.0

# ---- Testing -----------------------------------------------
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-mock>=3.12.0
httpx>=0.27.0              # For async test client

# ---- Utilities ---------------------------------------------
tenacity>=8.2.3            # Retry with backoff
httpx>=0.27.0
orjson>=3.9.15             # Fast JSON
rich>=13.7.0               # Pretty terminal output
click>=8.1.7               # CLI helpers
